tests/wpt/web-platform-tests/html/canvas/element/transformations/2d.transformation.translate.3d.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.transformation.translate.3d</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.transformation.translate.3d</h1>
<p class="desc">translate() function with three arguments modifies the underlying matrix appropriately</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("translate() function with three arguments modifies the underlying matrix appropriately");
_addTest(function(canvas, ctx) {

const dx = 2;
const dy = 3;
const dz = 4;
ctx.translate(dx, dy, dz);
let canvasTransform = ctx.getTransform();
_assert(canvasTransform.m41 = dx, "canvasTransform.m41 = dx");
_assert(canvasTransform.m42 = dy, "canvasTransform.m42 = dy");
_assert(canvasTransform.m43 = dz, "canvasTransform.m43 = dz");
ctx.translate(dx, dy, dz);
canvasTransform = ctx.getTransform();
_assert(canvasTransform.m41 = 2 * dx, "canvasTransform.m41 = 2 * dx");
_assert(canvasTransform.m42 = 2 * dy, "canvasTransform.m42 = 2 * dy");
_assert(canvasTransform.m43 = 2 * dz, "canvasTransform.m43 = 2 * dz");


});
</script>
